requirements.txt
pyyaml
semver
docopt
